add function 1차 완료
